tests/article.service.test.ts
import mongoose from 'mongoose';
// import { ArticleModel } from '../models/article';
import { ArticleService } from '../services/article.service';

describe('article service test', () => {
    // const article1 = { _id: '1', title: 'Mocked title 1', body: 'Mocked body 1', author: 'Mocked author 1', };
    // const article2 = { _id: '2', title: 'Mocked title 2', body: 'Mocked body 2', author: 'Mocked author 2', };

    beforeEach(() => {
        jest.useFakeTimers(); // this was added to attempt to fix 'find with id' test
    })

    test('test', () => {
        expect(ArticleService.test()).toEqual('OK!');
    })
    ,test('fetch with then', () => {
        ArticleService.fetch().then(result => {
            expect(jest.spyOn(ArticleService, 'find')).toHaveBeenCalledTimes(1);
            expect(result).resolves.toBeTruthy();
        });
    })

    // //this fails
    // ,test('fetch with async await', async () => {
    //     const result = await ArticleService.fetch();
    //     expect(jest.spyOn(ArticleService, 'find')).toHaveBeenCalledTimes(1);
    //     expect(result).resolves.toBeTruthy();
    // })

    // //this thows a ReferenceError
    // ,test('find with id', () => {
    //     const id = '1';
    //     ArticleService.find(id).then(result => {
    //         expect(jest.spyOn(ArticleService, 'find')).toBeCalledWith({_id: id});
    //         expect(result).resolves.toBeTruthy();
    //     })
    // })
})
